Gate Satellite Refinement (Workable now)
Refined the setup, Gate Sat ability will be activated at about 8000 range, the satellite has a 500 bombing range so it has to get close to the planet. Lowered the satellite armor, hull and shields considerably so when not engaged in a siege it will be vulnerable. Only works on enemy planets, neutral and home can't be targeted. Shierd regen and Shield mittigation are quite high during siege. There may be more but this is getting kinda long;)