Delete duplicate textures
Non-gui files that have dds duplicates.  Sins will use the dds files, so these appear to be unused and taking up space in the repo.  A backup of these files is on Google Drive.  We'll also need them if we want to update the images as we'll want to use the higher quality source.