Gate Satellite Ability Changed (Consumes a small amout of resources)
When the Gate Satellite is in siege mode, the energy fed through the gate from Asura in order to sustain its incredible shield regeneration and shield mitigation causes additional resource drain. The value is -0.05 (Sins displays this only as -0.1 but the underlying value should be correct) in addition to the usual ship upkeep. This drain is removed as soon as the satellite ability timer runs out (or it enters hyperspace) and the ship is no longer in range of an Enemy planet.