Fixed Missing Pact Strings
All pact strings are back in the .str file. It's very likely that we will come across this issue again when new factions are added. The strings are hardcoded (Sins expects the unique Pact research strings already defined in vanilla). If we wanted to be efficient, we would need to remove all our renamed IDS strings for those pacts and would link these hardcoded ones in. Atm, we have two sets of strings for Pacts where only one is necessary.