Asuran Epheus Titan - Lagrangian Beam damage and string fix
Fixed the missing strings for the Epheus Titan Lagrangian Beam ability. When the Matter Bridge is activated, this will now be reflected in the Lagrangian ability with increased damage (was reversed before) and the damage amount will be displayed on the info card in the correct location. The string has also been connected so it notes the ability (Matter Bridge) that is affecting the beam.