Attempt to fix chevron / stock menu
Sometimes when you click a bottom button or Sins is in a pending state - the menu buttons will switch from the cheveron to the default images.  I've noticed this several times and in particular on the Watch and Load icons.  I've had success in renaming our textures in such situations, so let's give it a try.