Reworked Time Dilation and Human Heracles Hero
The Time Dilation ability is now only available on the Human Heracles, no longer a Phoenix Cruiser ability. It requires a ZPM Artifact and research in order to work but has been buffed to lower many enemy stats when active, the effect intensifies with higher ability levels. The Heracles also lost its Asgard Beam ability to make room for Time Dilation but the Asgard Beams have been incorporated as weapon like on the Daedalus.