Titan research is now dependent on both Asgard Beams & Shield Research
This change should force the AI to research the Asgard Beam chain as well as the Shield chain which should make the Human AI a real challenge. Be aware that if this works as intended, the difference in difficulty should be quite pronounced.

Some Arrow errors have been introduced and we will eventually rearrange the research so that Sins can assign an existing arrow type.