Additional work on the Human Weapons Defense Platform
Abilities will be displayed only when researched, the Missiles Ability will only be available once research for the missile upgrade is completed. The Asgard Core Overload is now dependent on the Plasma Weapons research.

Added ability 2 points (Plasma Pulse and Asgard Core Overload) to the meshes and re-positioned the weapon 0 points to remove orientation errors.