Mark VIII nuke reworked: chance to hit
The Human Hero Nuke ability has been modified to include a chance to hit. If the nuke is diverted by the enemy, it'll end up detonating outside of the ship shields  causing 30k damage to shields and then hull if shields fail. If teleportation inside the ship is successful, 70k damage will be applied directly to the hull as well as 30k to shields and hull. Chance to penetrate shields with the nuke is 15% at level 1 and 25% at the highest level.

Cost in credits was lowered to 5000 (previously 10k)