Fix for Asuran Stargates (target filter)
The gates should now allow allied and player ships to use them, in fact anyone can use them once they are activated so choose locations wisely!

The starbase stargate was unable to function in gravity wells of stars which should be fixed now.

This is a temporary setup as stargates will eventually lose their phase stabilizer functionality and will function differently in general.