Quick Fix - Asuran Epheus Titan (shading/size)
The Asuran Epheus titan was reworked to be the intended size (resolve shield clipping).

The hardpoints are all new as well and now the Drone Frenzy will fire from all three towers simultaneously.

The model is still rough under the hood and the textures have issues. These will be fixed later.