Strings - Removed Duplicates, Added Missing References  #23
Identified that we have a lot of duplicate strings entries in the English.str.  This can make things confusing if we're updating strings, so I'm removing any duplicate IDs.  Also added strings that are referenced by entity files that don't exist.  Updated cosmetic, removing entities that will likely be removed in the future.