Texture Lightness Change & Disabled Tonemapping
Todd and Eternal were set to average match of the lightness boost & the tonemapping was disabled for a more natural look of the textures and to improve synchronization between texturing software and Sins PBR. The current texturing system textures are not meant to be used together with tonemapping, the result will be too heavy and intense.

PS: if anyone wants to have it active it's easy to do so, just remove the slashes in front of the definition.