Textures Update (Wraith Todd Hero & Siege Frigate)
Todd and the Siege Frigate were updated to v1.7.8. The textures should now work better with tonemapping so this will be enabled again as soon as the rest of these textures are updated which may happen later today.

Improvements to these textures include a fix for the height (normal map) issues for the inorganic material, color change more towards blue, removal of one of the effects that didn't work so well in this case and some roughness and metallic adjustments.